LPS-91461 Move menu item logic into helper class
